==> scripts/upload_sample_NCBI/scripts/01_sort_what_2_upload.Rmd <==
---
title: "Sort out which need to be uploaded"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
```

```{r}
all <- read_csv('/Volumes/vandenBrinkLab/Sarah/Project_auto/msk duke sdtable 01 18 20.csv') 
```

```{bash}
cat ../data/table_download_1.tsv ../data/table_download_2.tsv > ../data/current_on_ncbi_20200212.tsv
```

```{r}
current <- read_tsv('../data/current_on_ncbi_20200212.tsv')

current %>% 
  filter(BioSample.name  %in% new_samps) 
```
```{r}
# how many overlap 
length(intersect(current$BioSample.name, all$Sample.ID))

target <- all %>% 
  filter(!Sample.ID %in% current$BioSample.name)

already_upload <- current %>% 
  filter(BioSample.name %in% all$Sample.ID)

already_upload %>% 
  distinct(Submission)

already_upload %>% 
  write_csv('../data/nilu_samples_already_in_ncbi.csv')
```

```{r}
target %>% 
  write_csv('../data/Nilu_to_be_uploaded.csv')
```
